refactor: disable JIT by default
