+++ frugal_jit.go
@@ -1,4 +1,4 @@
-//go:build !go1.24 && amd64 && !windows
+//go:build frugal_jit
 
 /*
  * Copyright 2024 CloudWeGo Authors
@@ -20,17 +20,15 @@ package frugal
 
 import (
 	"reflect"
-	"sync"
 
+	"github.com/cloudwego/frugal/internal/jit"
 	"github.com/cloudwego/frugal/internal/jit/decoder"
 	"github.com/cloudwego/frugal/internal/jit/encoder"
-	"github.com/cloudwego/frugal/internal/jit/rt"
-	"github.com/cloudwego/frugal/internal/jit/utils"
 	"github.com/cloudwego/frugal/internal/opts"
 	"github.com/cloudwego/gopkg/protocol/thrift"
 )
 
-const jit = true
+const nojit = false
 
 func jitEncodedSize(val interface{}) int {
 	return encoder.EncodedSize(val)
@@ -44,79 +42,10 @@ func jitDecodeObject(buf []byte, val interface{}) (int, error) {
 	return decoder.DecodeObject(buf, val)
 }
 
-type _Ty struct {
-	d  int
-	ty *rt.GoType
-}
-
-var (
-	typool sync.Pool
-)
-
-func newty(ty *rt.GoType, d int) *_Ty {
-	if v := typool.Get(); v == nil {
-		return &_Ty{d, ty}
-	} else {
-		r := v.(*_Ty)
-		r.d, r.ty = d, ty
-		return r
-	}
-}
-
-// Pretouch compiles vt ahead-of-time to avoid JIT compilation on-the-fly, in
-// order to reduce the first-hit latency.
 func Pretouch(vt reflect.Type, options ...Option) error {
-	if opts.NoJIT {
-		return nil
-	}
-	d := 0
 	o := opts.GetDefaultOptions()
-
-	/* apply all the options */
 	for _, fn := range options {
 		fn(&o)
 	}
-
-	/* unpack the type */
-	v := make(map[*rt.GoType]bool)
-	t := rt.Dereference(rt.UnpackType(vt))
-
-	/* add the root type */
-	q := utils.NewQueue()
-	q.Enqueue(newty(t, 1))
-
-	/* BFS the type tree */
-	for !q.Empty() {
-		ty := q.Dequeue().(*_Ty)
-		tv, err := decoder.Pretouch(ty.ty, o)
-
-		/* also pretouch the encoder */
-		if err == nil {
-			err = encoder.Pretouch(ty.ty, o)
-		}
-
-		/* mark the type as been visited */
-		d, v[ty.ty] = ty.d, true
-		typool.Put(ty)
-
-		/* check for errors */
-		if err != nil {
-			return err
-		}
-
-		/* check for cutoff conditions */
-		if !o.CanPretouch(d) {
-			continue
-		}
-
-		/* add all the not visited sub-types */
-		for s := range tv {
-			if t = rt.UnpackType(s); !v[t] {
-				q.Enqueue(newty(t, d+1))
-			}
-		}
-	}
-
-	/* completed with no errors */
-	return nil
+	return jit.Pretouch(vt, o)
 }
